Revert "SE MANDAN DATOS PY A JS"
This reverts commit bcda6001e14a5f3827d73c1181e3f9c7570a7de2.

+++ public/stylesheets/javascripts/test.py
@@ -23,35 +23,21 @@ data['fecha'] = data['fecha'].apply(custom_date_parser)
 # Generar datos de la gráfica en el formato requerido por LightweightCharts
 charData = [{'time': date.strftime('%Y-%m-%d'), 'value': temp} for date, temp in zip(data['fecha'], data['temperatura'])]
 
-# Verificar si el archivo JavaScript existe y contiene datos en charData
-if os.path.exists('grafica4.js'):
+# Verificar si el archivo JavaScript existe y contiene datos
+if os.path.exists('grafica4.js') and os.path.getsize('grafica4.js') > 0:
+    # LEER EL CONTENIDO ACTUAL DEL ARCHIVO JAVASCRIPT
     with open('grafica4.js', 'r') as f:
-        lines = f.readlines()
-    
-    # Buscar la línea donde comienza la definición de charData
-    start_index = None
-    for i, line in enumerate(lines):
-        if 'var charData =' in line:
-            start_index = i
-            break
+        existing_code = f.read()
 
-    # Si se encontró la línea, eliminar todas las líneas de charData
-    if start_index is not None:
-        end_index = start_index
-        while end_index < len(lines) and '];' not in lines[end_index]:
-            end_index += 1
-        del lines[start_index:end_index+1]
-
-    # Escribir el archivo JavaScript con los datos actualizados
+    # ABRIR EL ARCHIVO JAVASCRIPT EN MODO ESCRITURA Y SOBRESCRIBIR EL NUEVO CODIGO
     with open('grafica4.js', 'w') as f:
         f.write("var charData = ")
         json.dump(charData, f)
         f.write(";\n")
-        f.writelines(lines)
+        f.write(existing_code)
 else:
-    # Si el archivo no existe, escribir los datos en un nuevo archivo JavaScript
+    # Escribir los datos en un nuevo archivo JavaScript
     with open('grafica4.js', 'w') as f:
-        f.write("var charData = ")
+        f.write("var chart_data = ")
         json.dump(charData, f)
         f.write(";\n")
-
